Allow usage of imported nodes
The existing implementation of the graphity! macro does not allow for
import of an external node. That was due to a limitation caused by
https://github.com/rust-lang/rust/issues/20400.

With this patch, the signature of the macro changes. It now accepts
direct reference to the node and its producer and consumer.

With this change, it is possible to use nodes defined in other crates.

Signed-off-by: Petr Horáček <phoracek@redhat.com>